Proyecto2/ReaderService/requirements.txt
google-cloud-core ==2.3.0
google-cloud-storage==2.3.0
pika=1.2.1